refactor: 버튼 위치 변경
